+++ src/app/component/examples/sync.route.component.ts
@@ -15,16 +15,16 @@ import { AComponent } from './AComponent';
             <sp-example-title [type]="exampleEnum.SYNC"></sp-example-title>
             <div class="split-example">
                 <as-split direction="vertical">
-                    <as-split-area size="30">
+                    <div as-split-area size="30">
                         <as-split direction="horizontal" (dragProgress)="update('b', $event)">
-                            <as-split-area [size]="sizes.a[0]">A 1</as-split-area>
+                            <div as-split-area [size]="sizes.a[0]">A 1</div>
                             <as-split-area [size]="sizes.a[1]">A 2</as-split-area>
                         </as-split>
-                    </as-split-area>
+                    </div>
                     <as-split-area size="70">
                         <as-split direction="horizontal" (dragProgress)="update('a', $event)">
                             <as-split-area [size]="sizes.b[0]">B 1</as-split-area>
-                            <as-split-area [size]="sizes.b[1]">B 2</as-split-area>
+                            <div as-split-area [size]="sizes.b[1]">B 2</div>
                         </as-split>
                     </as-split-area>
                 </as-split>